--- backend/src/main/java/com/woowacourse/moamoa/filter/exception/FilterNotExistException.java
@@ -1,8 +1,0 @@
-package com.woowacourse.moamoa.filter.exception;
-
-public class FilterNotExistException extends RuntimeException {
-
-    public FilterNotExistException() {
-        super("필터가 존재하지 않습니다.");
-    }
-}
