--- backend/src/main/java/com/woowacourse/moamoa/filter/service/response/FiltersResponse.java
@@ -1,12 +1,0 @@
-package com.woowacourse.moamoa.filter.service.response;
-
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class FiltersResponse {
-
-    private List<FilterResponse> tags;
-}
