--- backend/src/main/java/com/woowacourse/moamoa/study/domain/studyfilter/StudyFilter.java
@@ -1,34 +1,0 @@
-package com.woowacourse.moamoa.study.domain.studyfilter;
-
-import static javax.persistence.GenerationType.IDENTITY;
-import static lombok.AccessLevel.PROTECTED;
-
-import com.woowacourse.moamoa.filter.domain.Filter;
-import com.woowacourse.moamoa.study.domain.study.Study;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor(access = PROTECTED)
-public class StudyFilter {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "study_id")
-    private Study study;
-
-    @ManyToOne
-    @JoinColumn(name = "filter_id")
-    private Filter filter;
-}
