--- backend/src/main/java/com/woowacourse/moamoa/study/domain/studyfilter/StudySearchCondition.java
@@ -1,14 +1,0 @@
-package com.woowacourse.moamoa.study.domain.studyfilter;
-
-import com.woowacourse.moamoa.filter.domain.Filter;
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class StudySearchCondition {
-
-    private String title;
-    private List<Filter> filters;
-}
