+++ backend/src/main/java/com/woowacourse/moamoa/study/domain/studytag/StudySlice.java
@@ -1,0 +1,14 @@
+package com.woowacourse.moamoa.study.domain.studytag;
+
+import com.woowacourse.moamoa.study.domain.study.Study;
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class StudySlice {
+
+    private List<Study> studies;
+    private boolean hasNext;
+}
