+++ backend/src/main/java/com/woowacourse/moamoa/study/domain/studytag/StudyTag.java
@@ -1,0 +1,34 @@
+package com.woowacourse.moamoa.study.domain.studytag;
+
+import static javax.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PROTECTED;
+
+import com.woowacourse.moamoa.tag.domain.Tag;
+import com.woowacourse.moamoa.study.domain.study.Study;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor(access = PROTECTED)
+public class StudyTag {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "study_id")
+    private Study study;
+
+    @ManyToOne
+    @JoinColumn(name = "tag_id")
+    private Tag tag;
+}
