+++ backend/src/main/java/com/woowacourse/moamoa/tag/domain/repository/TagRepository.java
@@ -1,0 +1,11 @@
+package com.woowacourse.moamoa.tag.domain.repository;
+
+import com.woowacourse.moamoa.tag.domain.Tag;
+import java.util.List;
+import java.util.Optional;
+
+public interface TagRepository {
+
+    List<Tag> findAllByNameContainingIgnoreCase(String name);
+    Optional<Tag> findByName(String name);
+}
