+++ backend/src/main/java/com/woowacourse/moamoa/tag/service/response/TagsResponse.java
@@ -1,0 +1,12 @@
+package com.woowacourse.moamoa.tag.service.response;
+
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class TagsResponse {
+
+    private List<TagResponse> tags;
+}
